pyproject.toml
[project]
name = "device-policy-management-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "celery>=5.6.2",
    "django>=4.2.27",
    "django-restframework>=0.0.1",
    "djangorestframework>=3.16.1",
    "redis>=7.0.1",
]
